Classes relocated to org.hudsonci.test.ui package, added license, code clenup.
