--- hudson-test-ui/src/test/resources/test-context.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-                           http://www.springframework.org/schema/beans/spring-beans-2.5.xsd">
-
-    <bean id="propertyPlaceholderConfigurer"
-          class="org.springframework.beans.factory.config.PropertyPlaceholderConfigurer">
-        <property name="systemPropertiesModeName" value="SYSTEM_PROPERTIES_MODE_OVERRIDE"/>
-        <property name="properties">
-            <props>
-                <prop key="test.target.host">localhost</prop>
-                <prop key="test.target.port">6002</prop>
-                <prop key="test.target.contextroot">hudson</prop>
-            </props>
-        </property>
-    </bean>
-
-    <bean id="baseUrl" class="java.lang.String">
-        <constructor-arg value="http://${test.target.host}:${test.target.port}/${test.target.contextroot}"/>
-    </bean>
-</beans>
